rename consent_plus_data.js to consent_plus_insights.jst
