+++ lib/consent_plus_insights.js
